+++ tests/tests.rs
@@ -1,3 +1,5 @@
+#[macro_use]
+mod utils;
+
 mod buildtest;
 mod integration;
-mod utils;